Add gem 'rails-controller-testing' to Gemfile
There was a test inside the user's controller spec that looked for
the website to render the template for a new user that was failing.
The failure stated "assert_template has been extracted to a gem.
To continue using it, add `gem 'rails-controller-testing'` to your
Gemfile." The error was cleared once the specified gem was added
to the bottom of the Gemfile and bundle install was rerun.

- Add new gem 'rails-controller-testing' to Gemfile
- Rerun bundle install and rake spec