Create users controller spec file
https://trello.com/c/HJaXqsHD

Create the users controller spec file to test the different aspects
of the users controller. None of the aspects of the users controller
were being tested, therefore tests were inputted. There are currently
written and passing tests for the get #new action, which look for a
http status code of 200, assigning a new User instance to @user, and
rendering the :new template upon the users/new get request.

- Test for http status code 200
- Test for new instance of User to be @user
- Test for rending of users :new template