Add #new action for users
https://trello.com/c/HJaXqsHD

Add #new action to users controllers, because the action was needed to
pass information to the users/new route. The spec for the controller now
passes for #new action for users. Since the action cannot be empty there
is a @user variable set to a new instance of the User class.

- Define #new action in controller
- Set @user to new instance of User