Update users table in schema
The phone attribute for a User had the wrong input type of integer,
which needed to be changed to string to save correctly. Two
migrations were generated using rails to solve this problem. The
first migration removed the existing phone column from the users
table/class while the second re-added the phone column, but with
an input type of string. Once created, but files were migrated
and the schema successfully updated and will now work correctly.

- Generate and run migration for removing phone column
- Generate and run new migration for adding new phone column