Generate migration to add new phone column
The phone attribute for a User had the wrong input type of integer,
which needed to be changed to string to save correctly. The previous
migration was run to remove the pre-existing phone column with type
of integer, therefore this migration was run to add a new phone
column that uses string type. This was necessary since our user
validation require the presence of a phone number and the number
will be necessary for other aspects of the app.

- Generate migration for adding new phone column, type string