Create tests for POST #create
https://trello.com/c/HJaXqsHD

Since a new route and method were added to the controller, new
tests were necessary to check all aspects of the create user
action. There are two separate sections of tests, one for valid
parameters and the other for invalid parameters passed into the
User model. The valid parameter tests check for a proper http
status code of 302, a new user creation in the database, whether
they were assigned as a student or mentor, new user is automatically
logged in, then redirects to the users/index page. The invalid
parameters test for http status code 200, a new user to not be
added to the database but still assigned to @user, then re-renders
the new.hmtl template

- Add POST #create tests
- Two sections: valid & invalid parameters