Generate users index html template
The index html template was added to use for some of the new tests
that were written in the users controller spec. Since there was no
index hmtl to render, there were test errors. The template was only
created as a placeholder, it does not hold any html or code at all.

- Create empty index hmtl template for testing purposes