Move TravisCI yml to inside Rails
Upon closer inspection of our build fails on TravisCI's website I
noticed that all of the errors said that rake was aborted because
the site could not find a RAKEFILE, which was due to the fact that
we defined our original yml file outside of our actual Rails app.
Since the RAKEFILE only exists inside of the Rails app it was
automatically failing, therefore I moved the yml file to inside of
our app.

- Delete .travis.yml from Team-JTStream general repo
- Add .travis.yml to inside rails app dbc-mentor