Create model for Feedback
The users want to be able to leave feedback about their mentoring
session, whether they are a mentor or student. The feedback will be
sent to the opposite party, or type, of the user that fills it out.
Since the model did not exists and needs its own model, controller,
and routes, the model needed to be created.

- Create Feedback model