update appointment models and migrations to reflect needs
remove mentor_id since all creators of the appointments will be in a "mentor" role. add context to allow a mentor to see what the student needs to work on for an appointment. add tests for models and controller. add function to create appointments. add form view to create new appointment.