Generate Feedbacks table
To be able to save all of the feedback and its corresponding info-
rmation a table needed to be created in the database, which is why
a migration was generated and run for the app. The migration creates
all of the necessary feedback attributes such as mentor_id(integer),
student_id(integer), appointment_id(integer), written_by(string),
content(text), and rating(integer). The written_by attribute is used
to determine who to send the feedback to and when to display certain
types of feedback. The rating must be between 1-5 to determine an
overall score for a mentor/student to display on their profile page.

- Generate feedbacks migration
- Run feedbacks migration